plugin.h::MDP_Render_Info_t: Make width and height 32-bit.
Although only 16-bit is needed, 32-bit is faster with the asm functions.

+++ src/gens/plugins/plugin.h
@@ -79,8 +79,8 @@ typedef struct
 	
 	// Output screen buffer parameters.
 	void *screen;
-	unsigned short width;
-	unsigned short height;
+	unsigned int width;
+	unsigned int height;
 	int pitch;
 	int offset;
